Codigos de Clases/HF_Images/requirements.txt
fastapi
numpy
pydantic
opencv-python-headless
uvicorn[standard]
python-multipart
pillow